Move deleteOne function to module
