+++ src/deleteTodo.js
@@ -14,5 +14,21 @@ const deleteCompletedTodoItem = (tasks) => {
   });
 };
 
-export default deleteCompletedTodoItem;
+const deleteOnlyOne = (tasks) => {
+  const trashIcon = document.querySelectorAll('.fa-trash-alt');
+
+  trashIcon.forEach((icon) => icon.addEventListener('click', () => {
+    const { id } = icon.parentNode.id;
+
+    tasks.splice(id, 1);
+    for (let i = 0; i < tasks.length; i += 1) {
+      tasks[i].index = i;
+    }
+
+    localStorage.setItem('items', JSON.stringify(tasks));
+    location.reload();
+  }));
+};
+
+export { deleteCompletedTodoItem, deleteOnlyOne };
 /* eslint-enable no-restricted-globals */
